% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{cube.green}
\alias{cube.green}
\title{Cube Green}
\format{
\subsection{\code{cube.green}}{

An object of class \code{colour}.
}
}
\source{
\url{https://github.com/zerogetsamgow/cubecolours}
}
\usage{
cube.green
}
\description{
The Cube Group green is used in majority, bringing maturity and a seriousness
when needed.
}
\examples{
plot(cube.green)
}
\seealso{
\code{\link[=pal_colour]{pal_colour()}}
}
\keyword{datasets}
