% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{cube.grey}
\alias{cube.grey}
\title{Cube Grey}
\format{
\subsection{\code{cube.grey}}{

An object of class \code{colour}.
}
}
\usage{
cube.grey
}
\description{
The Cube Group grey is used as a neutral background colour.
}
\examples{
plot(cube.grey)
}
\keyword{datasets}
