% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{cube.lightorange}
\alias{cube.lightorange}
\title{Cube Pink}
\format{
\subsection{\code{cube.lightorange}}{

An object of class \code{colour}.
}
}
\usage{
cube.lightorange
}
\description{
The Cube Group pink, technically a light orange, adds warmth, livening up layouts and highlighting
key information.
}
\examples{
plot(cube.lightorange)
}
\keyword{datasets}
