% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{cube.pink}
\alias{cube.pink}
\title{Cube Pink}
\format{
\subsection{\code{cube.pink}}{

An object of class \code{colour}.
}
}
\usage{
cube.pink
}
\description{
The Cube Group pink adds warmth, livening up layouts and highlighting
key information.
}
\examples{
plot(cube.pink)
}
\keyword{datasets}
