% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{cube_colours}
\alias{cube_colours}
\title{Cube Colours}
\format{
\subsection{\code{cube_colours}}{

An object of class \code{palette} with 3 colours.
Use \code{names(cube_colours)} to return all colour names.
}
}
\source{
\url{https://github.com/zerogetsamgow/cubecolours}
}
\usage{
cube_colours
}
\description{
Cube Group's colour palette uses modern and distinctive tones
to convey corporate warmth.
}
\examples{
# Plot all colours
plot(cube_colours)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}#'
}
\keyword{datasets}
